+++ src/module/actor/pilot-sheet.ts
@@ -70,8 +70,8 @@ export class LancerPilotSheet extends ActorSheet {
   * The prepared data object contains both the actor data as well as additional sheet options
   */
   getData(): LancerPilotSheetData {
-    let data: LancerPilotSheetData = super.getData() as LancerPilotSheetData;
-    
+    const data: LancerPilotSheetData = super.getData() as LancerPilotSheetData;
+
     this._prepareItems(data);
     
     // Populate the callsign if blank (new Actor)
@@ -175,7 +175,7 @@ export class LancerPilotSheet extends ActorSheet {
     // Macro triggers
     if (this.actor.owner) {
       // Stat rollers
-      let statMacro = html.find(".stat-macro");
+      let statMacro = html.find(".roll-stat");
       statMacro.click((ev: any) => {
         ev.stopPropagation(); // Avoids triggering parent event handlers
         
@@ -246,6 +246,20 @@ export class LancerPilotSheet extends ActorSheet {
     if (!this.options.editable) return;
     
     if (this.actor.owner) {
+      // Customized increment/decrement arrows
+      let decr = html.find('button[class*="mod-minus-button"]');
+      decr.click((ev: any) => {
+        const but = $(ev.currentTarget);
+        but.next()[0].value = but.next()[0].valueAsNumber - 1;
+        this.submit({});
+      });
+      let incr = html.find('button[class*="mod-plus-button"]');
+      incr.click((ev: any) => {
+        const but = $(ev.currentTarget);
+        but.prev()[0].value = but.prev()[0].valueAsNumber + 1;
+        this.submit({});
+      });
+
       // Item/Macroable Dragging
       const statMacroHandler = (e: Event) => this._onDragMacroableStart(e);
       const talentMacroHandler = (e: DragEvent) => this._onDragTalentMacroableStart(e);
@@ -361,8 +375,8 @@ export class LancerPilotSheet extends ActorSheet {
         this.actor.update({ "data.mech_loadout.mounts": mounts });
         this._onSubmit(ev);
       });
-      
-      // Cloud upload
+
+      // Cloud download
       let download = html.find('.cloud-control[data-action*="download"]');
       download.click((ev: any) => {
         ev.stopPropagation();
@@ -459,7 +473,7 @@ export class LancerPilotSheet extends ActorSheet {
           );
           return;
         }
-        
+      
         if (item) {
           // Swap mech frame
           if (item.type === "frame") {
@@ -503,16 +517,16 @@ export class LancerPilotSheet extends ActorSheet {
               }
               
               let mount_element = $(event.target.closest(".lancer-mount-container"));
-              
+
               if (mount_element.length) {
                 let mount_whitelist = {
-                  Auxiliary: ["Integrated", "Aux-Aux", "Main", "Flex", "Main-Aux", "Heavy"],
-                  Main: ["Integrated", "Main", "Flex", "Main-Aux", "Heavy"],
-                  Heavy: ["Integrated", "Heavy"],
-                  Superheavy: ["Integrated", "Heavy"],
-                  Other: ["Integrated", "Aux-Aux", "Main", "Flex", "Main-Aux", "Heavy"],
+                  Auxiliary: [MountType.Integrated, MountType.Aux, MountType.AuxAux, MountType.MainAux, MountType.Flex, MountType.Main, MountType.Heavy],
+                  Main: [MountType.Integrated, MountType.Main, MountType.Flex, MountType.MainAux, MountType.Heavy],
+                  Heavy: [MountType.Integrated, MountType.Heavy],
+                  Superheavy: [MountType.Integrated, MountType.Heavy],
+                  Other: [MountType.Integrated, MountType.Aux, MountType.AuxAux, MountType.MainAux, MountType.Flex, MountType.Main, MountType.Heavy]
                 };
-                
+
                 let mount = mounts[parseInt(mount_element.data("itemId"))];
                 let valid = mount_whitelist[(item as LancerMechWeapon).data.data.mount];
                 if (!valid.includes(mount.type)) {
@@ -520,13 +534,13 @@ export class LancerPilotSheet extends ActorSheet {
                 } else if (item.data.data.mount === "Superheavy" && !mount.secondary_mount) {
                   ui.notifications.error(
                     "Assign a secondary mount to this heavy mount in order to equip a superheavy weapon"
-                    );
-                  } else {
-                    let weapon = await actor.createOwnedItem(duplicate(item.data));
-                    mount.weapons.push(weapon);
-                    console.log(`${lp} Inserted Mech Weapon into Mount`, weapon);
-                    this.actor.update({ "data.mech_loadout.mounts": mounts });
-                    this._onSubmit(event);
+                  );
+                } else {
+                  let weapon = await actor.createOwnedItem(duplicate(item.data));
+                  mount.weapons.push(weapon);
+                  console.log(`${lp} Inserted Mech Weapon into Mount`, weapon);
+                  this.actor.update({ "data.mech_loadout.mounts": mounts });
+                  this._onSubmit(event);
                   }
                 } else {
                   ui.notifications.error(